ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
joblib>=1.3.0
scikit-learn>=1.0.0
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
pypdf>=4.2.0
python-docx>=1.1.2
